requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
openai
weasyprint
markdown2
python-dotenv
